Move the particle system to anvil
