+++ Anvil/particle_system.cpp
@@ -8,20 +8,20 @@
 #include <array>
 #include <memory>
 
-namespace spkt {
+namespace anvil {
 namespace {
 
-ParticleSingleton& get_particle_runtime(spkt::registry& registry)
+spkt::ParticleSingleton& get_particle_runtime(spkt::registry& registry)
 {
-    auto entity = registry.find<ParticleSingleton>();
+    auto entity = registry.find<spkt::ParticleSingleton>();
     if (!registry.valid(entity)) [[unlikely]] {
         entity = registry.create();
-        registry.emplace<Runtime>(entity);
-        registry.emplace<NameComponent>(entity, "::ParticleRuntimeSingleton");
-        auto& ps = registry.emplace<ParticleSingleton>(entity);
-        ps.particles = std::make_shared<std::array<spkt::particle, NUM_PARTICLES>>();
+        registry.emplace<spkt::Runtime>(entity);
+        registry.emplace<spkt::NameComponent>(entity, "::ParticleRuntimeSingleton");
+        auto& ps = registry.emplace<spkt::ParticleSingleton>(entity);
+        ps.particles = std::make_shared<std::array<spkt::particle, spkt::NUM_PARTICLES>>();
     }
-    return registry.get<ParticleSingleton>(entity);
+    return registry.get<spkt::ParticleSingleton>(entity);
 }
 
 // Generates a random point in a sphere of radius R
@@ -52,7 +52,7 @@ void particle_system(spkt::registry& registry, double dt)
 
     // Go through all particle emitter entities and check to see if they need to emit
     // any particles this frame. If they do, add them to the particle array.
-    for (auto [pc, tc] : registry.view_get<ParticleComponent, Transform3DComponent>()) {
+    for (auto [pc, tc] : registry.view_get<spkt::ParticleComponent, spkt::Transform3DComponent>()) {
         pc.accumulator += (float)dt;
         while (pc.accumulator > pc.interval) {
             // Add a particle to the particle array
@@ -63,7 +63,7 @@ void particle_system(spkt::registry& registry, double dt)
                 .scale = pc.scale,
                 .life = pc.life
             };
-            ps.next_slot = --ps.next_slot % NUM_PARTICLES;
+            ps.next_slot = --ps.next_slot % spkt::NUM_PARTICLES;
 
             pc.accumulator -= pc.interval;
         }
