+++ Sprocket/CMakeLists.txt
@@ -47,7 +47,6 @@ add_library(Sprocket STATIC
             Scene/loader.cpp
             Scene/lua_ecs.cpp
 
-            Scene/Systems/particle_system.cpp
             Scene/Systems/physics_system.cpp
             Scene/Systems/basic_systems.cpp
             Scene/Systems/input_system.cpp
